First commit in experimental branch
I this branch I'm trying to implement MPI_NEIGHBOR_ALLTOALLW to perform global communication of the single component of velocity in one call. Future experimentation could be aimed to use that MPI subroutine (in conjunction with the definition of proper MPI derived data types) to communicate the entire velocity vector all at once.