Makefile edit; IN3 edit
Some options are added back to the makefile (the last one which I used and now produces a runtime error is -fcheck=all); IN3 file is modified accordingly to the TGV case. All this modifications are aimed at having a working code, but a lot of stuff should be put in the IN3 file and read by the program at runtime.